requirements.txt
aenum==3.1.11
amqp==5.1.1
anyio==3.6.1
apispec==5.2.2
apispec-webframeworks==0.5.2
appnope==0.1.3
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
backports.cached-property==1.0.1
billiard==3.6.4.0
boto3==1.22.6
botocore==1.25.6
Brotli==1.0.9
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
delayed-assert==0.3.6
Deprecated==1.2.13
etcd3==0.12.0
executing==0.8.3
falcon==3.1.0
falcon-apispec==0.4.0
falcon-swagger-ui==1.2.2
Flask==2.1.2
fonttools==4.33.3
greenlet==1.1.2
grpcio==1.46.1
grpcio-tools==1.46.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hyperframe==6.0.1
idna==3.3
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
kiwisolver==1.4.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.3
overrides==6.1.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
playwright==1.22.0
polling==0.3.2
prompt-toolkit==3.0.29
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyee==8.1.0
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.1
reportportal-client==5.2.2
requests==2.27.1
rfc3986==1.5.0
robotframework==5.0.1
robotframework-assertion-engine==0.5.1
robotframework-browser==12.4.0
robotframework-pythonlibcore==3.0.0
robotframework-reportportal==5.2.1
robotframework-requests==0.9.3
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
stack-data==0.2.0
tenacity==8.0.1
threadpoolctl==3.1.0
traitlets==5.2.0
typing-utils==0.1.0
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.1.2
wrapt==1.14.1
